refactor: extract market movers into tools/market/movers.py (#574)
Move get_top_movers_sp500, get_top_100, get_top_movers, and
get_stocks_by_tag into a focused market/movers.py module.
The original robinhood_market_data_tools.py re-exports them
for backward compatibility. MCP tool names unchanged.
