+++ tests/unit/test_market_research_tools.py
@@ -5,6 +5,7 @@ from unittest.mock import AsyncMock, patch
 
 import pytest
 
+from open_stocks_mcp.tools.market.movers import get_top_movers_sp500
 from open_stocks_mcp.tools.robinhood_market_data_tools import (
     get_stock_earnings,
     get_stock_events,
@@ -12,7 +13,6 @@ from open_stocks_mcp.tools.robinhood_market_data_tools import (
     get_stock_news,
     get_stock_ratings,
     get_stock_splits,
-    get_top_movers_sp500,
 )
 
 
@@ -21,9 +21,9 @@ class TestTopMoversSP500:
 
     @pytest.mark.journey_research
     @pytest.mark.unit
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.execute_with_retry")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_session_manager")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_rate_limiter")
+    @patch("open_stocks_mcp.tools.market.movers.execute_with_retry")
+    @patch("open_stocks_mcp.tools.market.movers.get_session_manager")
+    @patch("open_stocks_mcp.tools.market.movers.get_rate_limiter")
     @pytest.mark.journey_research
     @pytest.mark.unit
     @pytest.mark.asyncio
@@ -78,9 +78,9 @@ class TestTopMoversSP500:
 
     @pytest.mark.journey_research
     @pytest.mark.unit
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.execute_with_retry")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_session_manager")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_rate_limiter")
+    @patch("open_stocks_mcp.tools.market.movers.execute_with_retry")
+    @patch("open_stocks_mcp.tools.market.movers.get_session_manager")
+    @patch("open_stocks_mcp.tools.market.movers.get_rate_limiter")
     @pytest.mark.journey_research
     @pytest.mark.unit
     @pytest.mark.asyncio
@@ -138,9 +138,9 @@ class TestTopMoversSP500:
     @pytest.mark.skip(reason="Slow exception test - run with pytest -m exception_test")
     @pytest.mark.journey_research
     @pytest.mark.unit
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.execute_with_retry")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_session_manager")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_rate_limiter")
+    @patch("open_stocks_mcp.tools.market.movers.execute_with_retry")
+    @patch("open_stocks_mcp.tools.market.movers.get_session_manager")
+    @patch("open_stocks_mcp.tools.market.movers.get_rate_limiter")
     @pytest.mark.journey_research
     @pytest.mark.unit
     @pytest.mark.asyncio
@@ -170,7 +170,7 @@ class TestTopMoversSP500:
     @pytest.mark.skip(reason="Slow exception test - run with pytest -m exception_test")
     @pytest.mark.journey_research
     @pytest.mark.unit
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_session_manager")
+    @patch("open_stocks_mcp.tools.market.movers.get_session_manager")
     @pytest.mark.journey_research
     @pytest.mark.unit
     @pytest.mark.asyncio
